move tiling logic to pre/postprocessing. use pre/postprocessor in benchmarking
