+++ src/lightly_train/_task_models/ltdetr_instance_segmentation/task_model.py
@@ -32,6 +32,10 @@ from lightly_train._export.onnx_helpers import (
 from lightly_train._models import package_helpers
 from lightly_train._models.ecvit.ecvit import ECViTModelWrapper
 from lightly_train._models.ecvit.ecvit_package import EDGE_CRAFTER_PACKAGE
+from lightly_train._pre_post_processing import (
+    instance_segmentation as instance_segmentation_utils,
+)
+from lightly_train._pre_post_processing import tiling
 from lightly_train._task_models.instance_segmentation_components.edgecrafter_decoder import (
     ECSegTransformer,
 )
@@ -45,7 +49,6 @@ from lightly_train._task_models.ltdetr_instance_segmentation.config import (
 from lightly_train._task_models.ltdetr_object_detection.ecvit_vit_wrapper import (
     ECViTBackboneWrapper,
 )
-from lightly_train._task_models.object_detection_components import tiling_utils
 from lightly_train._task_models.object_detection_components.hybrid_encoder import (
     HybridEncoder,
 )
@@ -825,7 +828,7 @@ class LTDETRInstanceSegmentation(TaskModel):
             threshold=threshold,
         )[0]
 
-        tiles, coordinates = tiling_utils.tile_image(
+        tiles, coordinates = tiling.tile_image(
             image=x,
             overlap=overlap,
             tile_size=self.image_size,
@@ -893,19 +896,21 @@ class LTDETRInstanceSegmentation(TaskModel):
             )
             scores_tiles = torch.empty(0, dtype=dtype, device=device)
 
-        labels, masks, scores = tiling_utils.combine_instance_segmentation_tiles(
-            pred_global={
-                "labels": pred_global["labels"],
-                "masks": pred_global["masks"],
-                "scores": pred_global["scores"],
-            },
-            pred_tiles={
-                "labels": labels_tiles,
-                "masks": masks_tiles,
-                "scores": scores_tiles,
-            },
-            nms_iou_threshold=nms_iou_threshold,
-            global_local_iou_threshold=global_local_iou_threshold,
+        labels, masks, scores = (
+            instance_segmentation_utils.combine_instance_segmentation_tiles(
+                pred_global={
+                    "labels": pred_global["labels"],
+                    "masks": pred_global["masks"],
+                    "scores": pred_global["scores"],
+                },
+                pred_tiles={
+                    "labels": labels_tiles,
+                    "masks": masks_tiles,
+                    "scores": scores_tiles,
+                },
+                nms_iou_threshold=nms_iou_threshold,
+                global_local_iou_threshold=global_local_iou_threshold,
+            )
         )
 
         # Drop instances whose binarized mask has no positive pixels. Such masks
