+++ src/lightly_train/_task_models/object_detection_components/utils.py
@@ -10,11 +10,7 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.#
-"""Copyright(c) 2023 lyuwenyu. All Rights Reserved.
-
-# Modifications Copyright 2025 Lightly AG:
-- Added _yolo_to_xyxy and _denormalize_xyxy_boxes functions
-"""
+"""Copyright(c) 2023 lyuwenyu. All Rights Reserved."""
 
 from __future__ import annotations
 
@@ -291,43 +287,3 @@ def get_activation(act: str, inpace: bool = True):
         m.inplace = inpace
 
     return m
-
-
-def _yolo_to_xyxy(batch_boxes: list[Tensor]) -> list[Tensor]:
-    """Convert bounding boxes from YOLO (normalized cx, cy, w, h) format to
-    (normalized x_min, y_min, x_max, y_max) format.
-
-    Args:
-        boxes: Bounding boxes in YOLO format of shape (n_boxes, 4) with values
-            normalized between 0 and 1.
-
-    Returns:
-        Bounding boxes in (normalized x_min, y_min, x_max, y_max) format.
-    """
-    converted_boxes = []
-    for sample_boxes in batch_boxes:
-        cxcywh = sample_boxes
-        if cxcywh.ndim == 1:
-            cxcywh = cxcywh.reshape(-1, 4)
-        x_min = cxcywh[:, 0] - cxcywh[:, 2] / 2
-        y_min = cxcywh[:, 1] - cxcywh[:, 3] / 2
-        x_max = cxcywh[:, 0] + cxcywh[:, 2] / 2
-        y_max = cxcywh[:, 1] + cxcywh[:, 3] / 2
-        converted_boxes.append(torch.stack([x_min, y_min, x_max, y_max], dim=-1))
-    return converted_boxes
-
-
-def _denormalize_xyxy_boxes(
-    boxes: list[Tensor],
-    sizes: list[tuple[int, int]],
-) -> list[Tensor]:
-    """De-normalize bounding boxes from (normalized x_min, y_min, x_max, y_max) format."""
-    denormalized_boxes = []
-    for sample_boxes, (width, height) in zip(boxes, sizes):
-        sample_boxes_denormalized = sample_boxes.clone()
-        sample_boxes_denormalized[:, 0] = sample_boxes[:, 0] * width
-        sample_boxes_denormalized[:, 1] = sample_boxes[:, 1] * height
-        sample_boxes_denormalized[:, 2] = sample_boxes[:, 2] * width
-        sample_boxes_denormalized[:, 3] = sample_boxes[:, 3] * height
-        denormalized_boxes.append(sample_boxes_denormalized)
-    return denormalized_boxes
